Refactored CfApplicationCloner
